1. 讲BaseEntity Id 拆出来, 保留到实体中 2. 修改Save To Create 名称变化 3. R 响应体结构变化, 之后还会变更 为定下来 4. 自定义 Mapper 转换的一些东西 5. 将 getId 之类的关于Auth单独用Service 拆出来, 防止 鉴权框架耦合 5. 修改前端 Antd 的一些 CRUD 6. 目前 import 的一些 application , 后面看看用什么其他办法来加载, 目前不是很满意 7. (TODO)错误处理待完善 8. (TODO) 代码生成器 等待, 一个基础模板做完后在更新

+++ cuckoo-service/cuckoo-system-service/src/main/java/org/github/zhengchalei/service/impl/system/impl/SysUserServiceImpl.java
@@ -1,15 +1,19 @@
 package org.github.zhengchalei.service.impl.system.impl;
 
+import com.querydsl.core.BooleanBuilder;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.github.zhengchalei.entity.system.entity.SysRole;
 import org.github.zhengchalei.entity.system.entity.SysUser;
 import org.github.zhengchalei.entity.system.mapper.SysUserMapper;
-import org.github.zhengchalei.entity.system.request.SysUserSaveRequest;
+import org.github.zhengchalei.entity.system.request.SysUserCreateRequest;
 import org.github.zhengchalei.entity.system.request.SysUserUpdateRequest;
 import org.github.zhengchalei.entity.system.response.SysUserResponse;
 import org.github.zhengchalei.repository.system.SysRoleRepository;
 import org.github.zhengchalei.repository.system.SysUserRepository;
+import org.github.zhengchalei.service.impl.auth.impl.AuthService;
+import org.github.zhengchalei.service.impl.auth.impl.PasswordEncoderService;
 import org.github.zhengchalei.service.system.SysUserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,6 +24,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 import static org.github.zhengchalei.entity.system.entity.QSysUser.sysUser;
 
@@ -32,35 +37,53 @@ public class SysUserServiceImpl implements SysUserService {
     private final SysUserRepository sysUserRepository;
     private final SysUserMapper sysUserMapper;
     private final SysRoleRepository sysRoleRepository;
+    private final PasswordEncoderService passwordEncoderService;
+    private final AuthService authService;
 
     @Override
-    public Page<SysUserResponse> findSysUserPage(@NotNull Pageable pageable, @NotNull String username) {
-        Page<SysUser> page = sysUserRepository.findAll(sysUser.username.containsIgnoreCase(username), pageable);
+    public Page<SysUserResponse> findSysUserPage(@NotNull Pageable pageable, String username, Long roleId) {
+        BooleanBuilder predicate = new BooleanBuilder();
+        if (StringUtils.isNotBlank(username)) {
+            predicate.and(sysUser.username.containsIgnoreCase(username));
+        }
+        if (roleId != null) {
+            predicate.and(sysUser.roles.contains(new SysRole().setId(roleId)));
+        }
+        Page<SysUser> page = sysUserRepository.findAll(predicate, pageable);
         return page.map(sysUserMapper::toResponse);
     }
 
     @Override
     public List<SysUserResponse> findSysUserList(String username) {
-        return sysUserRepository.findAll().stream().map(sysUserMapper::toDto).collect(Collectors.toList());
+        BooleanBuilder predicate = new BooleanBuilder();
+        Iterable<SysUser> sysUsers = sysUserRepository.findAll(predicate);
+        return StreamSupport.stream(sysUsers.spliterator(), false).map(sysUserMapper::toResponse).collect(Collectors.toList());
     }
 
     @Override
-    public SysUserResponse findSysUserById(Long id) {
+    public SysUserResponse findSysUserById(@NotNull Long id) {
         return sysUserRepository.findById(id).map(sysUserMapper::toResponse).orElseThrow();
     }
 
     @Override
-    public SysUserResponse saveSysUser(SysUserSaveRequest sysUserSaveRequest) {
-        SysUser save = sysUserMapper.toEntity(sysUserSaveRequest);
+    public SysUserResponse createSysUser(@NotNull SysUserCreateRequest sysUserCreateRequest) {
+        SysUser save = sysUserMapper.toEntity(sysUserCreateRequest);
+        boolean exists = sysUserRepository.existsByUsernameIgnoreCaseAllIgnoreCase(save.getUsername());
+        if (exists) {
+            throw new RuntimeException("用户名已存在!");
+        }
+        // 密码加密
+        String password = passwordEncoderService.defaultPassword();
+        save.setPassword(password);
         // 查角色
-        Set<SysRole> roles = new HashSet<>(sysRoleRepository.findAllById(sysUserSaveRequest.getRoleIds()));
+        Set<SysRole> roles = new HashSet<>(sysRoleRepository.findAllById(sysUserCreateRequest.getRoleIds()));
         save.setRoles(roles);
         SysUser result = sysUserRepository.save(save);
         return sysUserMapper.toResponse(result);
     }
 
     @Override
-    public SysUserResponse updateSysUserById(Long id, SysUserUpdateRequest sysUserUpdateRequest) {
+    public SysUserResponse updateSysUserById(@NotNull Long id, @NotNull SysUserUpdateRequest sysUserUpdateRequest) {
         sysUserUpdateRequest.setId(id);
         SysUser flush = sysUserRepository.findById(id).orElseThrow();
         sysUserMapper.update(sysUserUpdateRequest, flush);
@@ -70,7 +93,7 @@ public class SysUserServiceImpl implements SysUserService {
     }
 
     @Override
-    public SysUserResponse partialSysUserById(Long id, SysUserUpdateRequest sysUserUpdateRequest) {
+    public SysUserResponse partialSysUserById(@NotNull Long id, @NotNull SysUserUpdateRequest sysUserUpdateRequest) {
         SysUser flush = sysUserRepository.findById(id).orElseThrow();
         sysUserMapper.partialUpdate(sysUserUpdateRequest, flush);
         Set<SysRole> roles = new HashSet<>(sysRoleRepository.findAllById(sysUserUpdateRequest.getRoleIds()));
@@ -79,7 +102,10 @@ public class SysUserServiceImpl implements SysUserService {
     }
 
     @Override
-    public void deleteSysUserById(Long id) {
+    public void deleteSysUserById(@NotNull Long id) {
+        if (authService.currentUser().getId().equals(id)) {
+            throw new RuntimeException("不能删除自己!");
+        }
         sysUserRepository.deleteById(id);
     }
 }
